+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 * Add new `radio_group` filter type, for any set of mutually exclusive options. The existing `boolean` filter now delegates to `radio_group` internally. ([#115](https://github.com/sevenseacat/cinder/issues/115))
 * Add `<:loading>`, `<:empty>`, and `<:error>` slots for custom state content, plus `error_message` attribute ([#20](https://github.com/sevenseacat/cinder/issues/20))
+* Support custom filter labels via `filter={[label: "..."]}` to override the column label in the filter UI ([#102](https://github.com/sevenseacat/cinder/issues/102))
 
 ### Bugfixes